parent 4069b95b3db63167622a3d9d8de213c0b4017f71 author Harrison Cramer <harrisoncramer@gmail.com> 1688136153 -0400 committer Harrison Cramer <harrisoncramer@gmail.com> 1691334576 -0400
Setup refactor and code cleanup

This MR makes several major tweaks to the codebase. Primarily it adjusts
the setup steps for the application so that rather than providing just
the project ID in the `.gitlab.nvim` file, users can also provide a
vareity of other settings, such as auth_token, base_branch, and so
forth. This is to make the project more extensible in the future.

This MR also fixes a variety of issues with error handling in the code,
primarily in the request/response model between the Lua jobs and the
Golang server.

BREAKING CHANGE: Modifies `.gitlab.nvim` and setup steps