Bugfix: Updates Diff Hashes
Rather than using branch names we are using the hashes provided directly
in the Gitlab API response, to compare the point at which the branch
diverged from the target to the head commit. We are additionally
flashing a warning if the MR contains a merge conflict.

This is a #patch change