Feat: Improve discussion tree toggling (#192)
* Fix: Toggle modified notes (#188)
* Fix: Toggle discussion nodes correctly
* Feat: Show Help keymap in discussion tree winbar

* Fix: Enable toggling nodes from the note body

Previously, the cursor had to be placed on the first line of the note
(the one with the author name) for toggling to work.

* Fix: Enable toggling resolved status from child nodes

Previously, the cursor had to be placed on the root node of the
discussion for toggling resolved status to work.

* Fix: Only try to show emoji popup on note nodes

Without this fix, there was an error whenever the cursor moved to a
file_name or path node in the "by_file_name" discussion tree.

* Feat: Add keymap for toggling tree type

Previously, tree type could only be set in the plugin configuration. Now
it can be toggled within a session. The keymap "i" is chosen based on a
similar functionality in DiffView.

* Fix: Disable tree type toggling in Notes

Tree type toggling only makes sense in the (linked) Discussions and
making it possible in the Notes could confuse users.