feat: add squash and remove source branch (#230)
This MR adds the ability to specify some settings when creating a merge request. Users may now specify that the original branch should be deleted after merge, and that the commits should be squashed into a single commit. It also updates the settings consistent with this change, and refactors numerous stylistic updates in the application.