Feat: Accept count with commenting keybindings
The keybindings `s` and `c` in the Reviewer windows now accept a count. It
works like this:

- `3sj` is identical to `s3j` - create a suggestion for the current and three
  more lines
- `3ss` or `s3s` - create a suggestion for three lines (including the current
  one)
- the same principal applies to `c` mappings

This is equivalent to how built-in keybindings like `3dj`, `d3j`, `3dd`, and
`d3d` work.

Unlike with builtin operators `d`, `c`, etc. the operator count and motion
count are NOT multiplied, so that the following key presses `3s2s` are not
equivalent to `6ss` but result in `s32s`.