fix: make dirty tree override imply_local properly
If `imply_local` is set but unused due to local changes present when review is started,
`imply_local` is overridden so that local changes do not unnecessarily block comments.

Now, uncommitted changes on a file will only block comments when they are made after the review is
started with `imply_local` applied.