fix: improve indication of resolved threads and drafts
The "resolved" checkmark is now directly after the number of resolved threads, e.g., "1✓/5". I find
this more readily comprehensible than "1/5✓".

The drafts are now shown *after* the unlinked comments - this makes the winbar text more stable when
adding / removing drafts, as the number of threads is placed always next to unlinked comments.